% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacedog.R
\name{spacedog}
\alias{spacedog}
\title{A Laika inspired ggplot theme. Laika, the first living creature to be shot into space and orbit Earth.}
\usage{
spacedog()
}
\description{
A Laika inspired ggplot theme. Laika, the first living creature to be shot into space and orbit Earth.
}
